macro scoping
