+++ utils/preferences/src/extensions/unicode/macros/struct_keyword.rs
@@ -4,7 +4,8 @@
 
 #[macro_export]
 /// TODO
-macro_rules! struct_keyword {
+#[doc(hidden)]
+macro_rules! __struct_keyword {
     ($(#[$doc:meta])* $name:ident, $ext_key:literal, $value:ty, $try_from:expr, $into:expr) => {
         #[derive(Debug, Clone, Eq, PartialEq)]
         #[allow(clippy::exhaustive_structs)] // TODO
@@ -27,7 +28,7 @@ macro_rules! struct_keyword {
             }
         }
 
-        impl $crate::preferences::PreferenceKey for $name {
+        impl $crate::PreferenceKey for $name {
             fn unicode_extension_key() -> Option<icu_locale_core::extensions::unicode::Key> {
                 Some(icu_locale_core::extensions::unicode::key!($ext_key))
             }
@@ -40,9 +41,11 @@ macro_rules! struct_keyword {
         }
     };
 }
+pub use __struct_keyword as struct_keyword;
 
 #[cfg(test)]
 mod tests {
+    use super::*;
     use core::str::FromStr;
     use icu_locale_core::{
         extensions::unicode,
