+++ utils/preferences/src/extensions/unicode/mod.rs
@@ -7,3 +7,4 @@
 pub mod errors;
 pub mod keywords;
 mod macros;
+pub use macros::*;
